// src/sp/types.ts
export interface ODataQueryable {
  expand?: string[];
  select?: string[];
}

export interface ODataQueryableCollection extends ODataQueryable {
  top?: number;
  orderBy?: string;
  orderyByAscending?: boolean;
  skip?: number;
  filter?: string;
}

export interface FilteredODataQueryable extends ODataQueryable {
  filter?: string;
}

export interface Scope {
  /** List GUID Id or title */
  list: string;

  /** Item Id */
  item?: number;
}

export interface PnpHookOptions<
  T = null | undefined | ODataQueryableCollection | ODataQueryable,
> extends ErrorOptions,
    RenderOptions,
    ContextOptions,
    BehaviourOptions {
  query?: T | undefined | null;
}